backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
httpx==0.25.1
python-dotenv==1.0.0
